Update blogpost.html
I hope the fourth one will be the beat. I tried with vi, code, note ++ and the last github editor. All of the previous I seeing the tag aligned but when I saw in the difference tool, I can't see aligned. Thanks for this challenge . If I won't see my user name, it just because the challenge was defeated